chore: productList type 위치 변경
