+++ src/app/product/_features/ProductList.tsx
@@ -1,16 +1,8 @@
 'use client';
 
-import type { ProductItem, ProductResponse } from '@/lib/types';
-import type { Ref } from 'react';
+import type { ProductResponse } from '@/lib/types';
 import ProductListView from './ProductListView';
 import useProductList from './useProductList';
-
-export type ProductListProps = {
-  isFetchingNextPage: boolean;
-  loadMoreRef: Ref<HTMLDivElement | null>;
-  handleCardButton: (item: ProductItem) => () => void;
-  products: ProductItem[];
-};
 
 export default function ProductList({
   initialProducts,
